types

+++ includes/seite_plugin.php
@@ -3,7 +3,7 @@
 use JTL\Plugin\Helper;
 use JTL\Shop;
 
-$kLink = (int)Shop::$kLink;
+$kLink = Shop::$kLink;
 if ($kLink !== null && $kLink > 0) {
     $linkFile   = Shop::Container()->getDB()->select('tpluginlinkdatei', 'kLink', $kLink);
     $linkHelper = Shop::Container()->getLinkService();
